requirements.txt
pandas
requests
python-dotenv
streamlit
folium
sqlalchemy
